Revert "promotion: preprod to prod"
