+++ build.gradle
@@ -251,7 +251,5 @@ implementation "com.stripe:stripe-java:28.0.1"
 implementation 'com.github.librepdf:openpdf:1.3.32'
 implementation 'io.github.openhtmltopdf:openhtmltopdf-pdfbox:1.1.31'
 implementation 'io.github.openhtmltopdf:openhtmltopdf-svg-support:1.1.31'
-implementation 'net.coobird:thumbnailator:0.4.14'
-
 
 }
