+++ src/main/java/app/bpartners/api/service/annotation/ExportAreaPictureAnnotationPDFGenerator.java
@@ -1,10 +1,10 @@
 package app.bpartners.api.service.annotation;
 
 import static app.bpartners.api.model.exception.ApiException.ExceptionType.SERVER_EXCEPTION;
-import static app.bpartners.api.service.annotation.factory.ExportAnnotationContextFactory.createContext;
 import static java.util.UUID.randomUUID;
 
 import app.bpartners.api.endpoint.rest.model.ExportAreaPictureAnnotation;
+import app.bpartners.api.endpoint.rest.model.ExportAreaPictureAnnotation3D;
 import app.bpartners.api.endpoint.rest.model.ExportAreaPictureAnnotationInstance;
 import app.bpartners.api.endpoint.rest.model.ExportAreaPictureAnnotationInstanceInfo;
 import app.bpartners.api.model.User;
@@ -20,6 +20,7 @@ import java.util.*;
 import lombok.SneakyThrows;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Component;
+import org.thymeleaf.context.Context;
 
 @Component
 public class ExportAreaPictureAnnotationPDFGenerator {
@@ -30,7 +31,8 @@ public class ExportAreaPictureAnnotationPDFGenerator {
   public static final String KEY_LABEL = "key";
   public static final String FONT_NAME = "Kumbh Sans";
   private static final String EMOJI_PATH = "fonts/twemoji/v/14.0.2/svg";
-  private static final String FONT_PATH = "fonts/KumbhSans-Regular.ttf";
+  private static final String BASE_64_URI_PREFIX = "data:image/png;base64,";
+  private static final String FONT_PATH = "fonts/KumbhSans-VariableFont_YOPQ,wght.ttf";
   private static final String AREA_PICTURE_ANNOTATION_TEMPLATE = "export-area-picture-annotations";
 
   public ExportAreaPictureAnnotationPDFGenerator(
@@ -59,21 +61,12 @@ public class ExportAreaPictureAnnotationPDFGenerator {
       builder.useFastMode();
       builder.withHtmlContent(html, null);
       builder.useSVGDrawer(new BatikSVGDrawer());
-      builder.useFont(
-          () -> {
-            try {
-              return new ClassPathResource(FONT_PATH).getInputStream();
-            } catch (IOException e) {
-              throw new RuntimeException(e);
-            }
-          },
-          FONT_NAME);
-
       builder.toStream(outputStream);
+      builder.useFont(new ClassPathResource(FONT_PATH).getFile(), FONT_NAME);
       builder.run();
 
       return outputStream.toByteArray();
-    } catch (RuntimeException | IOException e) {
+    } catch (RuntimeException e) {
       throw new ApiException(SERVER_EXCEPTION, e);
     }
   }
@@ -88,6 +81,101 @@ public class ExportAreaPictureAnnotationPDFGenerator {
 
     var context = createContext(user, logoBase64, annotation, annotationImages, annotation3DImages);
     return templateEngine.process(AREA_PICTURE_ANNOTATION_TEMPLATE, context);
+  }
+
+  private Context createContext(
+      User user,
+      String logoBase64,
+      ExportAreaPictureAnnotation annotation,
+      Pair<String, List<String>> annotationImages,
+      Pair<String, List<String>> annotation3DImages) {
+    var context = new Context();
+
+    var logoUri = base64ToUri(logoBase64);
+    var mainImageUri = base64ToUri(annotationImages.first());
+    var subImagesUris =
+        annotationImages.second().stream()
+            .map(ExportAreaPictureAnnotationPDFGenerator::base64ToUri)
+            .toList();
+
+    context.setVariable("user", user);
+    context.setVariable("userWebsite", user.getDefaultWebsite());
+    context.setVariable("logo", logoUri);
+    context.setVariable("address", annotation.getAddress());
+    context.setVariable("mainImage", mainImageUri);
+    context.setVariable("subImages", subImagesUris);
+    context.setVariable(
+        "pages", groupByFirstPage(GroupedByKey.from(annotation.getAnnotations()), 3, 3));
+
+    if (annotation.getLlm() != null) {
+      configureLLMContext(context, annotation);
+    }
+
+    if (annotation.getGlobalRateValue() != null || annotation.getGlobalRateType() != null) {
+      configureGlobalRateContext(context, annotation);
+    }
+    if (annotation.get3d() != null) {
+      configureAnnotation3DContext(context, annotation.get3d(), annotation3DImages);
+    }
+
+    return context;
+  }
+
+  private static void configureLLMContext(Context context, ExportAreaPictureAnnotation annotation) {
+    context.setVariable("llm", annotation.getLlm());
+  }
+
+  private static void configureGlobalRateContext(
+      Context context, ExportAreaPictureAnnotation annotation) {
+    context.setVariable("globalRateType", annotation.getGlobalRateType());
+    context.setVariable("globalRateValue", annotation.getGlobalRateValue());
+    context.setVariable(
+        "degradationLevels",
+        List.of(
+            Map.of("label", "A", "color", "#47BE62"),
+            Map.of("label", "B", "color", "#F4FBAB"),
+            Map.of("label", "C", "color", "#F9DD56"),
+            Map.of("label", "D", "color", "#F38F4B"),
+            Map.of("label", "E", "color", "#EF2C2D")));
+  }
+
+  private static void configureAnnotation3DContext(
+      Context context,
+      ExportAreaPictureAnnotation3D annotation3D,
+      Pair<String, List<String>> annotation3DImages) {
+    var pages = groupByFirstPage(annotation3D.getPans(), 3, 4);
+    var mainImage3DUri = base64ToUri(annotation3DImages.first());
+    var panImages3DUris =
+        annotation3DImages.second().stream()
+            .map(ExportAreaPictureAnnotationPDFGenerator::base64ToUri)
+            .toList();
+
+    context.setVariable("pages3D", pages);
+    context.setVariable("mainImage3D", mainImage3DUri);
+    context.setVariable("subImages3D", panImages3DUris);
+  }
+
+  public static <T> List<List<T>> groupByFirstPage(List<T> list, int firstPageMax, int limit) {
+    List<List<T>> pages = new ArrayList<>();
+    var iterator = list.iterator();
+
+    List<T> firstPage = new ArrayList<>();
+    for (int i = 0; i < firstPageMax && iterator.hasNext(); i++) {
+      firstPage.add(iterator.next());
+    }
+    if (!firstPage.isEmpty()) {
+      pages.add(firstPage);
+    }
+
+    while (iterator.hasNext()) {
+      List<T> page = new ArrayList<>();
+      for (int i = 0; i < limit && iterator.hasNext(); i++) {
+        page.add(iterator.next());
+      }
+      pages.add(page);
+    }
+
+    return pages;
   }
 
   public record GroupedByKey(String key, List<ExportAreaPictureAnnotationInstance> instances) {
@@ -135,6 +223,10 @@ public class ExportAreaPictureAnnotationPDFGenerator {
         .map(ExportAreaPictureAnnotationInstanceInfo::getValue)
         .findFirst()
         .orElse("generated_key_" + randomUUID());
+  }
+
+  private static String base64ToUri(String base64Image) {
+    return !base64Image.startsWith("data:") ? BASE_64_URI_PREFIX + base64Image : base64Image;
   }
 
   private EmojiReplacer getEmojiReplacer() {
