+++ src/main/resources/templates/export-area-picture-annotations.html
@@ -60,6 +60,7 @@
             width: calc(100% - 320px);
             vertical-align: top;
             word-wrap: break-word;
+            overflow-wrap: break-word;
             white-space: normal;
         }
 
@@ -89,6 +90,14 @@
         .annotation-item-title-value{ margin-right: 15px; }
         .annotation-item-info-label{ font-weight: bold; }
         .annotation-item-info-value{ font-weight: normal; }
+        .annotation-item-title-color {
+            display: none;
+            width: 1px;
+            height: 10px;
+            border-radius: 3px;
+            border: 10px solid;
+            margin-left: 5px;
+        }
         .annotation-item-info-item {
             font-size: 1rem; /* 12px */
             margin: 5px 0;
@@ -165,7 +174,7 @@
         }
 
         .llm section ul {
-            list-style-type: disc;
+            list-style-type: dot;
             padding-left: 0;
             margin: 0 auto;
             max-width: 100%;
@@ -177,6 +186,11 @@
             text-align: justify;
         }
 
+        .emoji {
+            display: inline-block;
+            line-height: 1;             /* size based on font-size */
+            vertical-align: -0.125em;   /* align baseline similar to text glyphs */
+        }
         .emoji svg,
         .emoji img {
             width: 1em;                 /* match text height */
